changed double xpBase in Creature to float

+++ BattleSim/src/main/java/com/revature/models/Creature.java
@@ -27,7 +27,7 @@ public class Creature {
 	private BattleClass battleClass;
 	
 	
-	private double xpBase;
+	private float xpBase;
 
 
 	public Creature() {
@@ -35,7 +35,7 @@ public class Creature {
 	}
 
 
-	public Creature(int crId, String name, Stats stats, BattleClass battleClass, double xpBase) {
+	public Creature(int crId, String name, Stats stats, BattleClass battleClass, float xpBase) {
 		super();
 		this.crId = crId;
 		this.name = name;
@@ -45,7 +45,7 @@ public class Creature {
 	}
 
 
-	public Creature(String name, Stats stats, BattleClass battleClass, double xpBase) {
+	public Creature(String name, Stats stats, BattleClass battleClass, float xpBase) {
 		super();
 		this.name = name;
 		this.stats = stats;
@@ -146,7 +146,7 @@ public class Creature {
 	}
 
 
-	public void setXpBase(double xpBase) {
+	public void setXpBase(float xpBase) {
 		this.xpBase = xpBase;
 	}
